supportive test function moved to a separate file
